refactor: move database tasks to background tasks
