+++ packages/maskbook/src/database/index.ts
@@ -1,5 +1,3 @@
-import './migrate/index'
-import './tasks/index'
 export * from './helpers/avatar'
 export * from './helpers/group'
 export * from './Persona/helpers'
