--- packages/maskbook/src/database/tasks/index.ts
@@ -1,7 +1,0 @@
-import { cleanProfileWithNoLinkedPersona } from './clean-avatar-profile'
-import { untilDocumentReady } from '../../utils/dom'
-import { sideEffect } from '../../utils/side-effects'
-
-sideEffect.then(untilDocumentReady).then(() => {
-    cleanProfileWithNoLinkedPersona()
-})
