+++ packages/maskbook/src/extension/background-script/Jobs/DatabaseCleanup/CleanProfileAndAvatar.ts
@@ -1,8 +1,9 @@
-import { createAvatarDBAccess, queryAvatarOutdatedDB, deleteAvatarsDB } from '../avatar'
-import { createTransaction } from '../helpers/openDB'
-import { consistentPersonaDBWriteAccess } from '../Persona/Persona.db'
-import { Identifier, ProfileIdentifier, GroupIdentifier } from '../type'
-import { IdentifierMap } from '../IdentifierMap'
+import { createAvatarDBAccess, queryAvatarOutdatedDB, deleteAvatarsDB } from '../../../../database/avatar'
+import { createTransaction } from '../../../../database/helpers/openDB'
+import { consistentPersonaDBWriteAccess } from '../../../../database/Persona/Persona.db'
+import { Identifier, ProfileIdentifier, GroupIdentifier } from '../../../../database/type'
+import { IdentifierMap } from '../../../../database/IdentifierMap'
+import { untilDocumentReady } from '../../../../utils'
 
 async function cleanAvatarDB(anotherList: IdentifierMap<ProfileIdentifier | GroupIdentifier, undefined>) {
     const t = createTransaction(await createAvatarDBAccess(), 'readwrite')('avatars', 'metadata')
@@ -13,8 +14,10 @@ async function cleanAvatarDB(anotherList: IdentifierMap<ProfileIdentifier | Grou
     await deleteAvatarsDB(Array.from(anotherList.keys()), t)
 }
 
-export async function cleanProfileWithNoLinkedPersona() {
-    setTimeout(cleanProfileWithNoLinkedPersona, 1000 * 60 * 60 * 24 * 7 /** days */)
+export default async function cleanProfileWithNoLinkedPersona(signal: AbortSignal) {
+    await untilDocumentReady()
+    const timeout = setTimeout(cleanProfileWithNoLinkedPersona, 1000 * 60 * 60 * 24 /** 1 day */)
+    signal.addEventListener('abort', () => clearTimeout(timeout))
 
     const cleanedList = new IdentifierMap<ProfileIdentifier | GroupIdentifier, undefined>(
         new Map(),
@@ -23,6 +26,7 @@ export async function cleanProfileWithNoLinkedPersona() {
     )
     const expired = new Date(Date.now() - 1000 * 60 * 60 * 24 * 14 /** days */)
     await consistentPersonaDBWriteAccess(async (t) => {
+        if (signal.aborted) throw new Error('')
         for await (const x of t.objectStore('profiles')) {
             if (x.value.linkedPersona) continue
             if (expired < x.value.updatedAt) continue
