+++ packages/maskbook/src/extension/background-script/Jobs/MigrateDatabase/index.ts
@@ -1,8 +1,7 @@
 import createUserGroupForOldUsers from './create.user.group.for.old.users'
 import fixQrPrivateKeyBug from './fix.qr.private.key.bug'
 import migratePeopleToPersona from './people.to.persona'
-import { untilDocumentReady } from '../../utils/dom'
-import { sideEffect } from '../../utils/side-effects'
+import { sideEffect, untilDocumentReady } from '../../../../utils'
 
 sideEffect.then(untilDocumentReady).then(run)
 function run() {
