+++ packages/maskbook/src/extension/background-script/Jobs/index.ts
@@ -1,3 +1,4 @@
 import './index_hmr'
 // Add and execute your non-cancelable jobs here. It won't compatible to HMR.
 import './PrintBuildFlags'
+import './MigrateDatabase'
