+++ packages/maskbook/src/extension/background-script/Jobs/index_hmr.ts
@@ -4,6 +4,7 @@ import * as InjectContentScripts from './InjectContentScripts'
 import * as NewInstalled from './NewInstalled'
 import * as PluginWorker from './StartPluginWorker'
 import * as SettingListeners from './SettingListeners'
+import * as CleanupProfileDatabase from './DatabaseCleanup/CleanProfileAndAvatar'
 
 type CancelableJob = { default: (signal: AbortSignal) => void }
 const CancelableJobs: CancelableJob[] = [
@@ -13,6 +14,7 @@ const CancelableJobs: CancelableJob[] = [
     IsolatedDashboardBridge,
     PluginWorker,
     SettingListeners,
+    CleanupProfileDatabase,
 ]
 
 const abort = new AbortController()
